Modified SQL Script
contain_* tables (for linking questions with lessons) were removed. Now
each question has a a foreign key which represent the related lesson.
In the answer table, now id_student and id_question is a unique
combination.
Also removed the choise answers / question tables. We first have to
decide if we are going to implement them.